MA-6791 Modify ethernet score to increase priority
As the wifi AP may have a 100 score when the device
boot up without ethernet connection and it may take over
the Ethernet. But once the Ethernet enabled before Wifi
the Ethernet will be OK.

So we modify the ethernet's score to 150 that will always
be larger than 100 even it is unvalidated which will be
minus 40 (110).

Change-Id: Ia7a9273f4a980560605970123c6ef053a138c1ea
Signed-off-by: Haoran Wang <b50027@freescale.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/456
Reviewed-by: Wang Haoran <elven.wang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/827
Reviewed-by: <yang.tian@nxp.com>